% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims_at.R
\name{claims_at}
\alias{claims_at}
\title{Claims At}
\usage{
claims_at(claimvaluations, valuationDate = NULL, claimAge = NULL,
  policyAge = NULL, maxValuationDate = NULL, dropNAs = FALSE)
}
\arguments{
\item{dropNAs}{}
}
\description{
Get claims evaluated at a specific date or age
}
\details{
Get claims evaluated at a specific date or age
}
\examples{
library(data.table)
}
