% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claimz}
\alias{claimz}
\title{Claims sample dataset}
\format{A data.table with 3 rows and 5 variables:
\describe{
  \item{ClaimID}{Unique identifier}
  \item{PolicyID}{Associated Policy ID}
  \item{EffectiveDate}{Effective date of the policy}
  \item{ExpirationDate}{Expiration date of the policy}
  \item{DOL}{Date of Loss}
}}
\usage{
claimz
}
\description{
A dataset with sample claims
}
\details{
library(data.table)

claimz <- data.table(
  ClaimID = c(1,2,3),
  PolicyID = c(1,1,2),
  DOL = as.Date(c("2014-3-15", "2014-4-1", "2015-5-30"))
)
claimz <- policiez[claimz, on="PolicyID"]
setcolorder(claimz, c("ClaimID", "PolicyID", "EffectiveDate", "ExpirationDate", "DOL"))

save(claimz, file="data/claimz.rda")
}
\keyword{datasets}
