% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_claims.R
\name{clean_claims}
\alias{clean_claims}
\title{Clean claims}
\usage{
clean_claims(claims, colmap = NULL)
}
\arguments{
\item{claims}{data.table object}

\item{colmap}{a mapping of column names in the form of a named character vector, or NULL}
}
\description{
Adjust the column names of a data.table of claims so that the resulting column names match the required names for claimz
functions.
}
\details{
Returns a data.table object, a copy of claims with required column names.
}
\examples{
library(data.table)

# Sample claims
claimz
claims <- copy(claimz)
setnames(claims, c("ClaimID"), c("claim_id"))
colmap <- c(
  "ClaimID"="claim_id",
  "PolicyID"="PolicyID",
  "EffectiveDate"="EffectiveDate",
  "ExpirationDate"="ExpirationDate",
  "DateOfLoss"="DateOfLoss",
  "ReportDate"="ReportDate"
)
clean_claims(claims, colmap)
}
