% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_claims.R
\name{is_valid_claims}
\alias{is_valid_claims}
\title{Is Valid claims}
\usage{
is_valid_claims(claims)
}
\arguments{
\item{claims}{data.table object with columns {"ClaimID", "PolicyID", "DateOfLoss"}}
}
\description{
Validate a claims dataset
}
\details{
Checks claims dataset for valid columns, values, and other properties.  Returns TRUE if valid, FALSE otherwise.
}
\examples{
library(data.table)

# Sample claims
claimz

# Validate
is_valid_claims(claimz)
is_valid_claims(claimz[, c("PolicyID", "EffectiveDate")])
}
