% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_claimvaluations.R
\name{is_valid_claimvaluations}
\alias{is_valid_claimvaluations}
\title{Is Valid claim-valuations}
\usage{
is_valid_claimvaluations(claimvaluations)
}
\arguments{
\item{claimvaluations}{data.table object with columns {"ValuationDate", "ClaimID"}}
}
\description{
Validate a claim-valuations dataset
}
\details{
Checks claim-valuations dataset for valid columns, values, and other properties.  Returns TRUE if valid, FALSE otherwise.
}
\examples{
library(data.table)

# Sample claimvaluations
claimvaluationz

# Validate
is_valid_claimvaluations(claimvaluationz)
is_valid_claimvaluations(claimvaluationz[, list(ClaimId = ClaimID, ValuationDate)])
}
