% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_datasets.R
\name{is_valid_datasets}
\alias{is_valid_datasets}
\title{Is Valid datasets}
\usage{
is_valid_datasets(policies, claims, claimvaluations)
}
\arguments{
\item{policies}{data.table object with columns {"PolicyID"}}

\item{claims}{data.table object with columns {"ClaimID", "PolicyID", "DateOfLoss"}}

\item{claimvaluations}{data.table object with columns {"ValuationDate", "ClaimID"}}
}
\description{
Validate policies, claims, and claimvaluations
}
\details{
Checks datasets for valid columns, values, and relational properties.  Returns TRUE if valid, FALSE otherwise.
}
\examples{
library(data.table)

# Sample datasets
policiez
claimz
claimvaluationz

# Validate
is_valid_datasets(policiez, claimz, claimvaluationz)
is_valid_datasets(policiez[1], claimz, claimvaluationz)
}
