% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_policies.R
\name{is_valid_policies}
\alias{is_valid_policies}
\title{Is Valid policies}
\usage{
is_valid_policies(policies)
}
\arguments{
\item{policies}{data.table object with columns {"PolicyID"}}
}
\description{
Validate a policies dataset
}
\details{
Checks policies dataset for valid columns, values, and other properties.  Returns TRUE if valid, FALSE otherwise.
}
\examples{
library(data.table)

# Sample policies
policiez

# Validate
is_valid_policies(policiez)
is_valid_policies(rbind(policiez, policiez))
}
