% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_triangles.R
\name{make_triangles}
\alias{make_triangles}
\title{Make Triangles}
\usage{
make_triangles(claimvaluations, format = "triangular", minLeftOrigin = NULL,
  originLength = 12, rowDev = 12, colDev = 12, lastValuationDate = NULL,
  fromMinLeftOrigin = TRUE, initialAge = originLength,
  colsFinancial = "auto", colmap_claimvaluations = NULL, verbose = FALSE)
}
\arguments{
\item{claimvaluations}{A data.table of cumulative transaction valuations (result of calling cumulate_transactions())}

\item{format}{How should the triangles be returned? Either "tall" (a data.table) or "triangular" (a list of data.tables)}

\item{minLeftOrigin}{See ?triangle_skeleton}

\item{originLength}{See ?triangle_skeleton}

\item{rowDev}{See ?triangle_skeleton}

\item{colDev}{See ?triangle_skeleton}

\item{lastValuationDate}{See ?triangle_skeleton}

\item{fromMinLeftOrigin}{See ?triangle_skeleton}

\item{initialAge}{See ?triangle_skeleton}

\item{colsFinancial}{What financial columns in \code{claimvaluations} should generate triangles? Default="auto" guesses}

\item{colmap_claimvaluations}{named vector of column mapping}

\item{verbose}{Should progress details be displayed?}
}
\description{
Build triangles
}
\details{
Returns a list of data.table objects
}
\examples{
library(data.table)

make_triangles(claimvaluationz, originLength = 3, rowDev = 3, colDev = 3)  # guess the financial columns
}
