% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policiez}
\alias{policiez}
\title{Policies sample dataset}
\format{A data.table with 2 rows and 3 variables:
\describe{
  \item{PolicyID}{Unique identifier}
  \item{EffectiveDate}{Effective date}
  \item{ExpirationDate}{Expiration date}
}}
\usage{
policiez
}
\description{
A dataset with sample policies
}
\details{
library(data.table)

policiez <- data.table::data.table(
  PolicyID = c(1,2),
  EffectiveDate = as.Date(c("2014-1-1", "2014-6-1")),
  ExpirationDate = as.Date(c("2015-1-1", "2015-6-1"))
)

save(policiez, file="data/policiez.rda")
}
\keyword{datasets}
