% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_losses.R
\name{simulate_losses}
\alias{simulate_losses}
\title{Simulate Losses}
\usage{
simulate_losses(numLosses = 500, startDate = as.Date("2010-01-01"),
  endDate = as.Date("2016-12-31"), cutoffDate = endDate)
}
\arguments{
\item{numLosses}{number of losses to simulate}

\item{startDate}{first date for which a loss can occur}

\item{endDate}{last date for which a loss can occur}

\item{cutoffDate}{last date for which a payment can occur}
}
\description{
Simulate a collection of claims and their Paid amounts over time
}
\details{
Returns a data.table with {ClaimID, ValuationDate, Paid, ...}.
}
\examples{
library(data.table)

set.seed(2017)
losses <- simulate_losses(numLosses = 500, startDate = as.Date("2010-01-01"), endDate = as.Date("2016-12-31"))
losses  # view losses

make_triangles(losses)
}
