% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies_at.R
\name{policies_at}
\alias{policies_at}
\title{Get policies evaluated at a specific date or age}
\usage{
policies_at(policies, claims, claimvaluations, valuationDate = NULL,
  policyAge = NULL, maxValuationDate = NULL, dropNAs = FALSE)
}
\description{
Get policies evaluated at a specific date or age
}

